refactor(ui): extract feed/toggles.ts — FeedViewToggle, FeedToggle, TagChip leaf components
